creating the test for the queue
